+++ NEWS.md
@@ -17,10 +17,14 @@
   * Added optional parameter `init_cohort$init_cohort_age` to specify a corresponding 
   initial age distribution in years (default = 0). Default ensures backwards compatibility.
   Age affects reproduction through `params_species$matural_age`
-  * Added optional parameter to `init_cohort`, to enable specification of initial
+  * Added optional parameter to `init_cohort`, to enable specification of initial plant
   nitrogen pools. Namely: `init_cohort_bl_n14`, `init_cohort_br_n14`,
   `init_cohort_bsw_n14`,`init_cohort_bHW_n14`,`init_cohort_seedC_n14`,
   `init_cohort_nsc_n14``. If not provided, default values ensure backwards compatibility.
+  * Added optional parameter to `init_soil`, to enable specification of initial soil
+  nitrogen pools. Namely: `init_fast_soil_N`, `init_slow_soil_N`, `init_pmicr_C`, 
+  `init_pmicr_d13C`, `init_pmicr_N`, `init_wcl1`, `init_wcl2`, `init_wcl3`, `init_N0_ecosystem`. 
+  If not provided, default values ensure backwards compatibility.
   * Added check that all `species$LMA` >= `params_tile$LMAmin`
 * P-model:
   * no changes